+++ pref/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Use flutter_lints
 - Actually update the state when a preference is changed
 - Add PrefDisabler
+- Add PrefChevron button
 
 ## 2.4.0